refactor: delete outdated files
