--- packages/superset-ui-preset-chart-xy/src/encodeable/types/Data.ts
@@ -1,5 +1,0 @@
-export type PlainObject<Key extends string = string, Value extends any = any> = {
-  [key in Key]: Value;
-};
-
-export type Dataset<T extends string = string> = Partial<PlainObject<T>>[];
