--- packages/superset-ui-preset-chart-xy/src/encodeable/types/Legend.ts
@@ -1,5 +1,0 @@
-export type Legend = boolean | null;
-
-export interface WithLegend {
-  legend?: Legend;
-}
