--- packages/superset-ui-preset-chart-xy/src/encodeable/utils/identity.ts
@@ -1,3 +1,0 @@
-const identity = (x: any) => x;
-
-export default identity;
