--- packages/superset-ui-preset-chart-xy/src/encodeable/utils/isDisabled.ts
@@ -1,11 +1,0 @@
-export default function isDisabled(
-  config:
-    | {
-        [key: string]: any;
-      }
-    | boolean
-    | null
-    | undefined,
-) {
-  return config === false || config === null;
-}
