--- packages/superset-ui-preset-chart-xy/src/encodeable/utils/isEnabled.ts
@@ -1,13 +1,0 @@
-import isDisabled from './isDisabled';
-
-export default function isEnabled(
-  config:
-    | {
-        [key: string]: any;
-      }
-    | boolean
-    | null
-    | undefined,
-) {
-  return !isDisabled(config);
-}
