+++ packages/superset-ui-preset-chart-xy/src/utils/constants.ts
@@ -1,2 +1,2 @@
-// eslint-disable-next-line import/prefer-default-export
+// // eslint-disable-next-line import/prefer-default-export
 export const DEFAULT_LABEL_ANGLE = 40;
